fix(argentina_pledge): Change mint amount and TC ID type to u64, fix storage type mismatch for Owner - audit recommendations LOW-2, LOW-4, INFO-2

+++ soroban/argentina_pledge/src/approval.rs
@@ -3,7 +3,7 @@ use crate::storage_types::DataKey;
 use crate::storage_types::{ApprovalAll, ApprovalKey};
 use soroban_sdk::{panic_with_error, Address, Env};
 
-pub fn read_approval(e: &Env, id: i128) -> Address {
+pub fn read_approval(e: &Env, id: u64) -> Address {
     let key = DataKey::Approval(ApprovalKey::ID(id));
     if let Some(approval) = e.storage().instance().get::<DataKey, Address>(&key) {
         approval
@@ -21,7 +21,7 @@ pub fn read_approval_all(e: &Env, owner: Address, operator: Address) -> bool {
     }
 }
 
-pub fn write_approval(e: &Env, id: i128, operator: Option<Address>) {
+pub fn write_approval(e: &Env, id: u64, operator: Option<Address>) {
     let key = DataKey::Approval(ApprovalKey::ID(id));
     match operator {
         Some(operator) => e.storage().instance().set(&key, &operator),
